requirements.txt
Pillow
django==1.10
django-material
django-filter==0.15.3
django-extensions
djangorestframework
django-wysiwyg-redactor
markdown
psycopg2
django-rest-swagger
django-cors-headers
django-rest-authemail
dj-database-url
newrelic
celery
awesome-slugify
git+git://github.com/eliangidoni/pycloudinary.git#egg=cloudinary
raven
